# docs/swagger.yaml
definitions:
  middle.ResCode:
    enum:
    - 200
    - 201
    - 202
    - 203
    - 204
    - 205
    - 206
    - 207
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeInvalidParam
    - CodeUserExist
    - CodeUserNotExist
    - CodeInvalidPassword
    - CodeServerBusy
    - CodeNeedLogin
    - CodeInvalidToken
  middle.ResponseData:
    properties:
      code:
        $ref: '#/definitions/middle.ResCode'
      data: {}
      msg: {}
    type: object
...
  service.HttpPath:
    properties:
      path:
        type: string
      pathType:
        $ref: '#/definitions/v1.PathType'
      serviceName:
        type: string
      servicePort:
        type: integer
    type: object
...
  service.WorkflowCreate:
    properties:
      container_port:
        type: integer
      cpu:
        type: string
      health_check:
        type: boolean
      health_path:
        type: string
      hosts:
        additionalProperties:
          items:
            $ref: '#/definitions/service.HttpPath'
          type: array
        type: object
      image:
        type: string
      label:
        additionalProperties:
          type: string
        type: object
      memery:
        type: string
      name:
        type: string
      namespace:
        type: string
      node_port:
        type: integer
      port:
        type: integer
      replicas:
        type: integer
      type:
        type: string
    type: object
...
  v1.PathType:
    enum:
    - Exact
    - Prefix
    - ImplementationSpecific
    type: string
    x-enum-varnames:
    - PathTypeExact
    - PathTypePrefix
    - PathTypeImplementationSpecific
info:
  contact: {}
paths:
...
  /api/k8s/workflow/create:
    post:
      consumes:
      - application/json
      description: 创建workflow
      operationId: /api/k8s/workflow/create
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.WorkflowCreate'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "创建成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 创建workflow
      tags:
      - Workflow
  /api/k8s/workflow/del:
    delete:
      consumes:
      - application/json
      description: 删除Workflow
      operationId: /api/k8s/workflow/del
      parameters:
      - description: Workflow ID
        in: query
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "删除成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 删除Workflow
      tags:
      - Workflow
  /api/k8s/workflow/id:
    get:
      consumes:
      - application/json
      description: 根据ID查看workflow
      operationId: /api/k8s/workflow/id
      parameters:
      - description: Workflow ID
        in: query
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": }'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 根据ID查看workflow
      tags:
      - Workflow
  /api/k8s/workflow/list:
    get:
      consumes:
      - application/json
      description: 查看Configmap列表
      operationId: /api/k8s/workflow/list
      parameters:
      - description: 过滤
        in: query
        name: filter_name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 分页限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": }'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 查看Configmap列表
      tags:
      - Workflow
...
swagger: "2.0"